Split widget rendering components by domain #300
